refactor: Move seed data outside of main seed file
